Support deltas on up to 2TB partitions
It looks like the xdelta3 library will need to be updated to support
higher values. This will address the need to support >4GB partitions
which are becoming more common.

+++ src/functions.h
@@ -66,7 +66,7 @@ struct fun_context {
     // Delta update processing (NULL if not in use)
     struct xdelta_state *xd;
     off_t xd_source_offset;
-    size_t xd_source_count;
+    off_t xd_source_count;
     const char *xd_source_path;
     struct disk_crypto *xd_source_dc;
